Move MessageStream into its own file
